requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/base.in
#
click==8.1.3
    # via -r requirements/base.in
docutils==0.18.1
    # via python-daemon
lockfile==0.12.2
    # via python-daemon
python-daemon==2.3.0
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
